feat(auth, user): restrict manage user to superadmin

+++ app/Http/Controllers/UserController.php
@@ -20,6 +20,7 @@ class UserController extends Controller
     public function index()
     {
         try {
+            $this->authorize('manage', 'manage-user');
             $users = $this->userServiceInterface->getAllUsers();
             return ApiResponse::success(UserResource::collection($users), 'Users Retrieved', 200);
         } catch (\Exception $e) {
@@ -30,6 +31,7 @@ class UserController extends Controller
     public function show(int $id)
     {
         try {
+            $this->authorize('manage', 'manage-user');
             $user = $this->userServiceInterface->getUserById($id);
             return ApiResponse::success(new UserResource($user), 'User Retrieved', 200);
         } catch (\Exception $e) {
@@ -40,6 +42,7 @@ class UserController extends Controller
     public function store(StoreUserRequest $request)
     {
         try {
+            $this->authorize('manage', 'manage-user');
             $user = $this->userServiceInterface->createUser($request->getStorePayload());
             return ApiResponse::success(new UserResource($user), 'User Create successsful', 201);
         } catch (\Exception $e) {
@@ -50,6 +53,7 @@ class UserController extends Controller
     public function update(UpdateUserRequest $request, int $id)
     {
         try {
+            $this->authorize('manage', 'manage-user');
             $this->userServiceInterface->updateUser($request->getUpdatePayload(), $id);
             return ApiResponse::success('', 'User Update successsful', 200);
         } catch (\Exception $e) {
@@ -60,6 +64,7 @@ class UserController extends Controller
     public function destroy(int $id)
     {
         try {
+            $this->authorize('manage', 'manage-user');
             $this->userServiceInterface->deleteUser($id);
             return ApiResponse::success('', 'User Delete successsful', 204);
         } catch (\Exception $e) {
